backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai-whisper==20231117
easyocr==1.7.1
numpy==1.26.3
Pillow==10.2.0
aiofiles==23.2.1
httpx==0.26.0